Moved class files and Deprecated methods in RC
