+++ app/src/main/AndroidManifest.xml
@@ -55,15 +55,16 @@
             android:value="0" />
 
         <activity
-            android:name=".SearchMapsActivity"
+            android:name=".Searching.SearchMapsActivity"
             android:label="@string/title_activity_search_maps" />
-        <activity android:name=".SearchActivity" />
+        <activity android:name=".Searching.SearchActivity" />
         <activity android:name=".HelpActivity" />
         <activity android:name=".RiderRequestActivity" />
         <activity
             android:name=".MakeRequestActivity"
             android:windowSoftInputMode="adjustPan" />
-        <activity android:name=".Notifications.NotificationActivity"></activity>
+        <activity android:name=".Notifications.NotificationActivity" />
+        <activity android:name=".Searching.SearchResultsActivity"></activity>
     </application>
 
 </manifest>
