+++ app/src/main/java/comcmput301f16t01/github/carrier/RequestController.java
@@ -8,21 +8,21 @@ import comcmput301f16t01.github.carrier.Notifications.Notification;
 import comcmput301f16t01.github.carrier.Notifications.NotificationController;
 
 /**
- * Singleton Pattern
- * We will be using RequestController as a middleman between the activity and elasticController.
- * * modifies/returns a RequestList model
- * * @see Request
- * * @see RequestList
+ * Uses a singleton pattern to query and get results of requests.
+ *
+ * Typical use:
+ *      rc.searchType( params );            // Change the singleton's information
+ *      requestList = rc.getResults();      // Get the results of that search (global to program)
+ *
+ * Or use one of the getX() functions to get immediate info.
  */
 public class RequestController {
-    private static ArrayList<Request> requestList = null;
+    private static ArrayList<Request> requestList;
     private Context saveContext = null;
-
-    //private static RequestList requestList = null;
-
-
+    
     /**
-     * Since we are using elastic search to get the request list.
+     * Prevents errors when a RequestController is initialized and methods that require requestList
+     * to not be null (i.e. getResult() )
      */
     public RequestController() {
         if (requestList == null) {
@@ -30,19 +30,7 @@ public class RequestController {
         }
     }
 
-    /**
-     * @return the RequestList held by this controller.
-     */
-    public static ArrayList<Request> getInstance() {
-        if (requestList == null) {
-            requestList = new ArrayList<Request>();
-        }
-        return requestList;
-    }
-
-    /**
-     * @param request
-     */
+    /** Adds a request to elastic search. */
     public String addRequest(Request request) {
         if(request.getStart() == null || request.getEnd() == null) {
             return "You must first select a start and end location";
@@ -55,21 +43,13 @@ public class RequestController {
         }
     }
 
-    /**
-     * Deprecated: should use new function that uses elastic search or FileIO (depending on
-     * connectivity), not singleton?
-     */
-    @Deprecated
-    public ArrayList<Request> getRequests(User rider) {
-        return requestList;
+    /** Clears information in the singleton, not exactly necessary */
+    // TODO check the necessity of this function.
+    public void clear() {
+        requestList = new RequestList();
     }
 
-    /**
-     *
-     */
-    public void reset() {
-    }
-
+    // TODO Why does this need a rider? You can cancel a request just knowing the request.
     public void cancelRequest(User rider, Request request) {
     }
 
@@ -110,41 +90,23 @@ public class RequestController {
     public void payForRequest(Request request) {
     }
 
-
     /**
-     * Is used to search open requests by a location submitted by a driver. The search results
-     * should return an ordered list with priority given to requests that begin in closest
-     * proximity to the driver-submitted location. Search results are limited to a range of
-     * locations that are within a reasonable distance to the driver-submitted location.
-     *
-     * @param location The location submitted by the user to query by
-     * @return Returns the search results
+     * Search requests by the keyword, will set it so the singleton contains the information for
+     * this query. Use getResults() to get the information.
+     * @param keyword
      */
-    public ArrayList<Request> searchByLocation(Location location) {
-        return new ArrayList<>();
+    public void searchByKeyword(String keyword) {
+        //ElasticRequestController.SearchByKeywordTask sbkt = new ElasticRequestController.SearchByKeywordTask();
+        //sbkt.execute(keyword)
+        //requestList = sbkt.get();
     }
 
     /**
-     * Is used to search open requests by a keyword query submitted by a driver. The query string
-     * should not be case-dependent. The search matches the query string to the request description
-     * to determine whether a request should be included in the search results or not.
-     *
-     * @param query
-     * @return Returns the search results
+     * Search requests by a location. This sets it so the singleton contains the information for
+     * this query. Use getResults() to get the information.
      */
-    // TODO consider having these search results sorted by location and limited to a location range
-    //   this will require us to compare against the driver's current location...how do we get that?
-    public ArrayList<Request> searchByKeyword(String query) {
-        return new ArrayList<>();
-    }
+    public void searchByLocation( /* location parameters? */ ) {
 
-    /**
-     * Is used to provide a driver with a list of all open requests.
-     *
-     * @return An array list of open requests.
-     */
-    public ArrayList<Request> getOpenRequests() {
-        return new ArrayList<Request>();
     }
 
     /**
@@ -158,33 +120,85 @@ public class RequestController {
         return new ArrayList<Request>();
     }
 
+
+
+    /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+     * * * * * * * * * * * * * * * * *    DEPRECATED FUNCTIONS   * * * * * * * * * * * * * * * * * *
+     * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
+
     /**
-     * Will be used to get a list of drivers that the user can choose to confirm as there driver
-     *
-     * @param request The request that is being queried for a list of drivers.
-     * @return An ArrayList of drivers that have offered to give a ride on the request.
+     * Deprecated: use the void getSearchByLocation
      */
+    @Deprecated
+    public ArrayList<Request> getSearchByLocation(Location location) {
+        return new ArrayList<>();
+    }
+
+    /**
+     * Deprecated: This is literally built into a Request. (see the ArrayList of offered drivers)
+     */
+    @Deprecated
     public ArrayList<User> getDrivers(Request request) {
         return new ArrayList<User>();
     }
 
     /**
-     * Provides a list of requests that are available for a driver to accept.
-     *
-     * @return an arraylist of requests that a driver is able to accept.
-     * @See getOpenRequests
+     * Deprecated: There are several other functions that do this. Also, try to only use
+     * getSearchByKeyword or getSearchByLocation
      */
+    @Deprecated
     public ArrayList<Request> getAvailableRequests() {
+        return new ArrayList<Request>();
+    }
+
+    /** Get the results of a searchByKeyword or a getSearchByLocation query. */
+    public RequestList getResult() {
+        return (RequestList) requestList;
+    }
+
+    /**
+     * Deprecated: Use getResult instead.
+     */
+    @Deprecated
+    public static ArrayList<Request> getInstance() {
+        if (requestList == null) {
+            requestList = new ArrayList<Request>();
+        }
+        return requestList;
+    }
+
+    /**
+     * Deprecated: There is no user story that says we need to modify a request description after it has been
+     * created
+     */
+    @Deprecated
+    public void setRequestDescription(Request request, String description) {
+    }
+
+    /**
+     * Deprecated: Only use getSearchByKeyword or getSearchByLocation?
+     */
+    @Deprecated
+    public ArrayList<Request> getOpenRequests() {
         return new ArrayList<Request>();
     }
 
 
     /**
-     * This is used to set the Request description.
-     *
-     * @param request     The request that we are modifying
-     * @param description The description that we are setting it to.
+     * Deprecated: use the void function instead (singleton changer) so that this can be used with
+     * the getResults() method
      */
-    public void setRequestDescription(Request request, String description) {
+    @Deprecated
+    public ArrayList<Request> getSearchByKeyword(String query) {
+        return new ArrayList<>();
+    }
+
+    /**
+     * Deprecated: should use new function that uses elastic search or FileIO (depending on
+     * connectivity), not singleton?
+     */
+    @Deprecated
+    public ArrayList<Request> getRequests(User rider) {
+        return requestList;
     }
 }
