+++ app/src/main/java/comcmput301f16t01/github/carrier/RequestList.java
@@ -1,8 +1,10 @@
 package comcmput301f16t01.github.carrier;
 
+import java.util.ArrayList;
+
 /**
  * Created by Ben on 2016-10-09.
  */
-public class RequestList {
+public class RequestList extends ArrayList<Request> {
 
 }
