+++ app/src/main/java/comcmput301f16t01/github/carrier/Searching/SearchMapsActivity.java
@@ -1,7 +1,9 @@
-package comcmput301f16t01.github.carrier;
+package comcmput301f16t01.github.carrier.Searching;
 
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
+
+import comcmput301f16t01.github.carrier.R;
 
 public class SearchMapsActivity extends FragmentActivity {
