+++ app/src/main/res/layout/activity_search_maps.xml
@@ -5,4 +5,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="comcmput301f16t01.github.carrier.SearchMapsActivity" />
+    tools:context="comcmput301f16t01.github.carrier.Searching.SearchMapsActivity" />
