+++ app/src/test/java/comcmput301f16t01/github/carrier/RequestsTests.java
@@ -13,7 +13,7 @@ public class RequestsTests {
     @After
     public void clean() {
         RequestController rc = new RequestController();
-        rc.reset();
+        rc.clear();
     }
 
     /**
